requirements.txt
discord.py==2.5.2
python-dotenv==1.1.0
pytube==7.0.16
requests==2.32.3
youtube-search==2.1.2
PyNaCl==1.5.0
asyncio==3.4.3
wavelink==3.4.1